refactor(player-profile): lift the renewal negotiation into a hook
Every presentational card in this directory was extracted long ago; the logic
never was. `PlayerProfile.tsx` was 1029 lines, and the single largest thing in
it had nothing to do with rendering a profile — it was a contract negotiation.

`useContractRenewalFlow` now owns the whole flow: the 13 `renewal*` state
atoms, the derived validity and status values, opening and closing the modal,
the live financial-impact projection effect, submitting an offer by hand, and
delegating the talks to the assistant manager. `PlayerProfile.tsx` 1029 → 797.

It follows the hook template this directory already consumes —
`useFreeAgentContractFlow` and `useTransferBidFlow` in `../transfers/` — so the
call site reads the same as the two flows sitting beside it.

One cleanup travels with the move. `handleDelegateRenewal` was a nested
early-return chain repeating the same six `setRenewal*` calls per branch, which
made it hard to see that the branches differ in only one thing that matters:
whether the negotiation is left open. That is now a table, and the difference
is a single `isTerminal` column. The `failed` row doubles as the fallback for
an unrecognised status, preserving the `else` it replaced — a bare map lookup
would have thrown instead.

Behaviour is unchanged, including two writes that are invisible today but kept
faithful rather than tidied: only an agreed deal clears the suggested terms
from an earlier counter-offer round, and only the two non-successful outcomes
surface the assistant's note.

Covered by the 13 renewal tests in `PlayerProfile.test.tsx`, two of which were
added in the previous commit precisely so the delegate branches were pinned
before being restructured.
