+++ src/components/playerProfile/PlayerProfile.tsx
@@ -8,7 +8,6 @@ import {
 import { PlayerData, GameStateData } from "../../store/gameStore";
 import { ArrowLeft } from "lucide-react";
 import { useTranslation } from "react-i18next";
-import { resolveBackendText } from "../../utils/backendI18n";
 import { resolveTranslatedErrorMessage } from "../../utils/errorMessage";
 import {
   clearContractExitIntent,
@@ -52,20 +51,10 @@ import PlayerProfileRecentMatchesCard, {
 import PlayerProfileRenewalModal from "./PlayerProfileRenewalModal";
 import PlayerProfileSeasonStatsCard from "./PlayerProfileSeasonStatsCard";
 import {
-  type DelegatedRenewalCaseData,
-  type DelegatedRenewalResponseData,
-  type NegotiationFeedbackData,
-  getRenewalStatusClassName,
-  getRenewalStatusMessage,
-  type RenewalProjectionData,
-  type RenewalResponseData,
-  type RenewalStatus,
-  shouldDisableRenewalSubmit,
-} from "./PlayerProfile.renewal";
-import {
   getScoutAvailability,
   type PlayerProfileScoutStatus,
 } from "./PlayerProfile.scouting";
+import { useContractRenewalFlow } from "./useContractRenewalFlow";
 
 interface PlayerProfileProps {
   player: PlayerData;
@@ -107,27 +96,6 @@ export default function PlayerProfile({
   const [scoutStatus, setScoutStatus] =
     useState<PlayerProfileScoutStatus>("idle");
   const [scoutError, setScoutError] = useState<string | null>(null);
-  const [showRenewalModal, setShowRenewalModal] = useState(false);
-  const [renewalWage, setRenewalWage] = useState("");
-  const [renewalLength, setRenewalLength] = useState("2");
-  const [renewalSubmitting, setRenewalSubmitting] = useState(false);
-  const [renewalStatus, setRenewalStatus] = useState<RenewalStatus>("idle");
-  const [renewalError, setRenewalError] = useState<string | null>(null);
-  const [renewalSuggestedWage, setRenewalSuggestedWage] = useState<
-    number | null
-  >(null);
-  const [renewalSuggestedYears, setRenewalSuggestedYears] = useState<
-    number | null
-  >(null);
-  const [renewalSessionStatus, setRenewalSessionStatus] =
-    useState<RenewalResponseData["session_status"]>("idle");
-  const [renewalIsTerminal, setRenewalIsTerminal] = useState(false);
-  const [renewalCooledOff, setRenewalCooledOff] = useState(false);
-  const [renewalFeedback, setRenewalFeedback] =
-    useState<NegotiationFeedbackData | null>(null);
-  const [renewalProjection, setRenewalProjection] = useState<
-    RenewalProjectionData["projection"] | null
-  >(null);
   const [contractActionSubmitting, setContractActionSubmitting] =
     useState(false);
   const [contractActionError, setContractActionError] = useState<string | null>(
@@ -176,34 +144,6 @@ export default function PlayerProfile({
       : contractRiskLevel === "warning"
         ? t("finances.contractRiskWarning")
         : t("finances.contractRiskStable");
-  const renewalOfferedWage = Number(renewalWage);
-  const renewalOfferedYears = Number(renewalLength);
-  const isRenewalWageValid =
-    Number.isFinite(renewalOfferedWage) && renewalOfferedWage > 0;
-  const isRenewalLengthValid =
-    Number.isInteger(renewalOfferedYears) && renewalOfferedYears > 0;
-  const renewalViolatesSoftCap =
-    isRenewalWageValid &&
-    renewalProjection !== null &&
-    !renewalProjection.policy_allows;
-  const renewalSubmitDisabled = shouldDisableRenewalSubmit({
-    renewalSubmitting,
-    renewalIsTerminal,
-    isRenewalWageValid,
-    isRenewalLengthValid,
-    renewalViolatesSoftCap,
-  });
-  const renewalStatusMessage = getRenewalStatusMessage(
-    {
-      renewalSessionStatus,
-      renewalStatus,
-      renewalSuggestedWage,
-      renewalSuggestedYears,
-      renewalError,
-    },
-    t,
-  );
-  const renewalStatusClassName = getRenewalStatusClassName(renewalStatus);
   const scoutAvailability = getScoutAvailability({
     staff: gameState.staff,
     scoutingAssignments: gameState.scouting_assignments || [],
@@ -277,41 +217,32 @@ export default function PlayerProfile({
     handleMakeBid,
   } = useTransferBidFlow({ gameState, onGameUpdate });
 
-  function openRenewalModal(): void {
-    setRenewalWage(String(player.wage));
-    setRenewalLength("2");
-    setRenewalSubmitting(false);
-    setRenewalStatus("idle");
-    setRenewalError(null);
-    setRenewalSuggestedWage(null);
-    setRenewalSuggestedYears(null);
-    setRenewalSessionStatus("idle");
-    setRenewalIsTerminal(false);
-    setRenewalCooledOff(false);
-    setRenewalFeedback(null);
-    setRenewalProjection(null);
-
-    const renewalState = player.morale_core?.renewal_state;
-    const blockedUntil = renewalState?.manager_blocked_until;
-    const hasActiveManagerBlock =
-      renewalState?.status === "Blocked" &&
-      (!blockedUntil ||
-        blockedUntil.slice(0, 10) >= gameState.clock.current_date.slice(0, 10));
-    if (hasActiveManagerBlock) {
-      setRenewalSessionStatus("blocked");
-      setRenewalIsTerminal(true);
-    }
-
-    setShowRenewalModal(true);
-  }
-
-  function closeRenewalModal(): void {
-    if (renewalSubmitting) {
-      return;
-    }
-
-    setShowRenewalModal(false);
-  }
+  const {
+    showRenewalModal,
+    renewalWage,
+    setRenewalWage,
+    renewalLength,
+    setRenewalLength,
+    renewalSubmitting,
+    renewalIsTerminal,
+    renewalCooledOff,
+    renewalFeedback,
+    renewalProjection,
+    isRenewalWageValid,
+    renewalViolatesSoftCap,
+    renewalSubmitDisabled,
+    renewalStatusMessage,
+    renewalStatusClassName,
+    openRenewalModal,
+    closeRenewalModal,
+    handleRenewalSubmit,
+    handleDelegateRenewal,
+  } = useContractRenewalFlow({
+    player,
+    gameState,
+    hasAssistantManager,
+    onGameUpdate,
+  });
 
   useEffect(() => {
     setHasConsumedInitialRenewalIntent(false);
@@ -355,41 +286,6 @@ export default function PlayerProfile({
     showTerminationModal,
     startWithTerminationModal,
   ]);
-
-  useEffect(() => {
-    if (!showRenewalModal || !isRenewalWageValid) {
-      setRenewalProjection(null);
-      return;
-    }
-
-    let cancelled = false;
-
-    const loadProjection = async (): Promise<void> => {
-      try {
-        const result = await invoke<RenewalProjectionData>(
-          "preview_renewal_financial_impact",
-          {
-            playerId: player.id,
-            weeklyWage: renewalOfferedWage,
-          },
-        );
-
-        if (!cancelled) {
-          setRenewalProjection(result.projection ?? null);
-        }
-      } catch {
-        if (!cancelled) {
-          setRenewalProjection(null);
-        }
-      }
-    };
-
-    loadProjection();
-
-    return () => {
-      cancelled = true;
-    };
-  }, [isRenewalWageValid, player.id, renewalOfferedWage, showRenewalModal]);
 
   useEffect(() => {
     let cancelled = false;
@@ -483,145 +379,6 @@ export default function PlayerProfile({
       cancelled = true;
     };
   }, [player.id, player.stats.appearances]);
-
-  async function handleRenewalSubmit(): Promise<void> {
-    if (renewalSubmitDisabled) {
-      return;
-    }
-
-    setRenewalSubmitting(true);
-    setRenewalStatus("idle");
-    setRenewalError(null);
-    setRenewalCooledOff(false);
-
-    try {
-      const result = await invoke<RenewalResponseData>("propose_renewal", {
-        playerId: player.id,
-        weeklyWage: renewalOfferedWage,
-        contractYears: renewalOfferedYears,
-      });
-
-      onGameUpdate?.(result.game);
-      setRenewalStatus(result.outcome);
-      setRenewalSuggestedWage(result.suggested_wage);
-      setRenewalSuggestedYears(result.suggested_years);
-      setRenewalSessionStatus(result.session_status);
-      setRenewalIsTerminal(result.is_terminal);
-      setRenewalCooledOff(result.cooled_off ?? false);
-      setRenewalFeedback(result.feedback ?? null);
-
-      if (result.session_status === "blocked") {
-        setRenewalStatus("blocked");
-      }
-
-      if (result.outcome === "counter_offer") {
-        if (result.suggested_wage !== null) {
-          setRenewalWage(String(result.suggested_wage));
-        }
-
-        if (result.suggested_years !== null) {
-          setRenewalLength(String(result.suggested_years));
-        }
-      }
-    } catch (error) {
-      setRenewalStatus("error");
-      setRenewalError(resolveTranslatedErrorMessage(error, t));
-      setRenewalCooledOff(false);
-    } finally {
-      setRenewalSubmitting(false);
-    }
-  }
-
-  async function handleDelegateRenewal(): Promise<void> {
-    if (renewalSubmitting) {
-      return;
-    }
-
-    if (!hasAssistantManager) {
-      setRenewalStatus("error");
-      setRenewalError(
-        resolveTranslatedErrorMessage(
-          "be.error.contracts.noAssistantManagerAssigned",
-          t,
-        ),
-      );
-      setRenewalCooledOff(false);
-      return;
-    }
-
-    setRenewalSubmitting(true);
-    setRenewalError(null);
-    setRenewalCooledOff(false);
-
-    try {
-      const result = await invoke<DelegatedRenewalResponseData>(
-        "delegate_renewals",
-        {
-          playerIds: [player.id],
-          maxWageIncreasePct: 35,
-          maxContractYears: 3,
-        },
-      );
-
-      onGameUpdate?.(result.game);
-      const delegatedCase: DelegatedRenewalCaseData | undefined =
-        result.report.cases.find(
-          (renewalCase) => renewalCase.player_id === player.id,
-        );
-
-      if (!delegatedCase) {
-        setRenewalStatus("error");
-        setRenewalError(t("playerProfile.renewalDelegateMissingReport"));
-        return;
-      }
-
-      if (delegatedCase.status === "successful") {
-        setRenewalStatus("accepted");
-        setRenewalSessionStatus("agreed");
-        setRenewalIsTerminal(true);
-        setRenewalSuggestedWage(null);
-        setRenewalSuggestedYears(null);
-        setRenewalCooledOff(false);
-        setRenewalFeedback(null);
-        return;
-      }
-
-      if (delegatedCase.status === "stalled") {
-        setRenewalStatus("rejected");
-        setRenewalSessionStatus("stalled");
-        setRenewalIsTerminal(false);
-        setRenewalCooledOff(false);
-        setRenewalFeedback(null);
-        setRenewalError(
-          resolveBackendText(
-            delegatedCase.note_key,
-            delegatedCase.note,
-            delegatedCase.note_params,
-          ),
-        );
-        return;
-      }
-
-      setRenewalStatus("blocked");
-      setRenewalSessionStatus("blocked");
-      setRenewalIsTerminal(true);
-      setRenewalCooledOff(false);
-      setRenewalFeedback(null);
-      setRenewalError(
-        resolveBackendText(
-          delegatedCase.note_key,
-          delegatedCase.note,
-          delegatedCase.note_params,
-        ),
-      );
-    } catch (error) {
-      setRenewalStatus("error");
-      setRenewalError(resolveTranslatedErrorMessage(error, t));
-      setRenewalCooledOff(false);
-    } finally {
-      setRenewalSubmitting(false);
-    }
-  }
 
   async function handleMarkLetExpire(): Promise<void> {
     if (contractActionSubmitting) {
